Broke out "MyAddPage" class to own file.
Clean up a lot of code, broke out widgets out of the build Widget.
