+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:erikstodoapp/secondView.dart';
 
 void main() {
   runApp(MaterialApp(
@@ -84,97 +85,71 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Things you gotta do, get to it!'),
-        centerTitle: true,
-        actions: <Widget>[
-          PopupMenuButton<String>(
-              onSelected: filterAction,
-              itemBuilder: (BuildContext context) {
-                return FilterOptions.choices
-                    .map((String choice) => PopupMenuItem<String>(
-                          value: choice,
-                          child: Text(choice),
-                        ))
-                    .toList();
-              })
-        ],
-      ),
-      body: ListView.builder(
-          itemCount: filteredToDoList.length,
-          itemBuilder: (context, index) {
-            return Card(
-              child: ListTile(
-                  title: Text(filteredToDoList[index].value),
-                  trailing: GestureDetector(
-                    child: Icon(Icons.remove_circle_outline),
-                    onTap: () {
-                      var itemToRemove = filteredToDoList[index];
-
-                      setState(() {
-                        filteredToDoList.remove(itemToRemove);
-                      });
-
-                      toDoList.removeWhere(
-                          (element) => element.id == itemToRemove.id);
-                    },
-                  ),
-                  leading: Checkbox(
-                    value: filteredToDoList[index].status,
-                    onChanged: (bool newValue) {
-                      setState(() {
-                        filteredToDoList[index].status = newValue;
-                      });
-                    },
-                  )),
-            );
-          }),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          moveToSecondPage();
-        },
-        splashColor: Colors.white,
-        child: Text(
-          '+',
-          style: TextStyle(fontSize: 30),
-        ),
-      ),
+          title: Text('Things you gotta do, get to it!'),
+          centerTitle: true,
+          actions: [
+            _filterMenu(),
+          ]),
+      body: _listBuilder(),
+      floatingActionButton: _addItemsToListButton(),
     );
   }
-}
 
-class MyAddPage extends StatefulWidget {
-  @override
-  _MyAddPageState createState() => _MyAddPageState();
-}
+  Widget _filterMenu() {
+    return PopupMenuButton<String>(
+        onSelected: filterAction,
+        itemBuilder: (BuildContext context) {
+          return FilterOptions.choices
+              .map((String choice) => PopupMenuItem<String>(
+                    value: choice,
+                    child: Text(choice),
+                  ))
+              .toList();
+        });
+  }
 
-class _MyAddPageState extends State<MyAddPage> {
-  final TextEditingController userInputController = new TextEditingController();
+  Widget _listBuilder() {
+    return ListView.builder(
+        itemCount: filteredToDoList.length,
+        itemBuilder: (context, index) {
+          return Card(
+            child: ListTile(
+                title: Text(filteredToDoList[index].value),
+                trailing: GestureDetector(
+                  child: Icon(Icons.cancel_outlined),
+                  onTap: () {
+                    var itemToRemove = filteredToDoList[index];
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: Text('Add something to your list!'),
-          centerTitle: true,
-        ),
-        body: Column(children: [
-          TextField(
-            controller: userInputController,
-            decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                hintText: 'Enter something to do'),
-          ),
-          RaisedButton(
-            onPressed: () {
-              if (userInputController.text != '' &&
-                  userInputController.text != null) {
-                Navigator.pop(context, userInputController.text);
-              }
-            },
-            child: const Text('Add', style: TextStyle(fontSize: 20)),
-            color: Colors.green,
-            splashColor: Colors.white,
-          ),
-        ]));
+                    setState(() {
+                      filteredToDoList.remove(itemToRemove);
+                    });
+
+                    toDoList.removeWhere(
+                        (element) => element.id == itemToRemove.id);
+                  },
+                ),
+                leading: Checkbox(
+                  value: filteredToDoList[index].status,
+                  onChanged: (bool newValue) {
+                    setState(() {
+                      filteredToDoList[index].status = newValue;
+                    });
+                  },
+                )),
+          );
+        });
+  }
+
+  Widget _addItemsToListButton() {
+    return FloatingActionButton(
+      onPressed: () {
+        moveToSecondPage();
+      },
+      splashColor: Colors.white,
+      child: Text(
+        '+',
+        style: TextStyle(fontSize: 30),
+      ),
+    );
   }
 }
